🐛 FIX: iOS heap corruption crash - SchedulerEvent OpaquePointer size mismatch
Root cause: typealias SchedulerEvent = OpaquePointer (8 bytes) vs C++ struct (16 bytes).
UnsafeMutablePointer<SchedulerEvent>.allocate() allocated N*8 bytes but rawEventDataToEvents
wrote N*16 bytes — a 2x buffer overflow corrupting nano malloc guard bytes.

Fix: Cast Dart's raw event bytes directly via UnsafeRawPointer.assumingMemoryBound() instead
of allocating an intermediate buffer. Also fixes position_frame_t (UInt64 → UInt32).

Includes: native C++ scheduler with lock-free render callbacks, fixed-size arrays replacing
unordered_map on audio thread, in-process AudioUnit loading for real-time safety.

+++ ios/Classes/SwiftCInterop.swift
@@ -7,7 +7,7 @@ import CoreAudio
 
 // Type aliases for C++ types
 typealias track_index_t = UInt32
-typealias position_frame_t = UInt64
+typealias position_frame_t = UInt32
 typealias Dart_Port = Int64
 typealias SchedulerEvent = OpaquePointer
 typealias SfizzDSPKernelAdapter = OpaquePointer
